Adding game loop functions in server with matching client functions; adding set entry tab in DB tool; updating firebase and jquery for tool; updating orb-board json file with new key info
