+++ client/gamefiles/css/main.css
@@ -69,6 +69,11 @@ header {
     font-size: 18px;
 }
 
+.text-button {
+    color: cornflowerblue;
+    cursor: pointer;
+}
+
 /* Action buttons */
 
 #actions {
@@ -143,6 +148,12 @@ header {
     color: white;
 }
 
+#board-content-header {
+    width: 100%;
+    height: 100px;
+    padding: 10px 0;
+}
+
 #board-start-button {
     display: none;
 }
@@ -154,12 +165,48 @@ header {
     font-weight: 600;
 }
 
+/* Board images */
+#board-worlds {
+    display: flex;
+    justify-content: space-between;
+
+}
+
+.world {
+    width: 135px;
+    height: 160px;
+    background-size: 135px 135px;
+    background-repeat: no-repeat;
+    text-align: center;
+}
+
+#world-earth {
+    background-image: url('../images/worlds/temp.png');
+}
+
+#world-luna {
+    background-image: url('../images/worlds/temp.png');
+}
+
+#world-mars {
+    background-image: url('../images/worlds/temp.png');
+}
+
+#world-venus {
+    background-image: url('../images/worlds/temp.png');
+}
+
+#world-mercury {
+    background-image: url('../images/worlds/temp.png');
+}
+
 /* Modals */
 
 #modal {
     display: none;
-    position: fixed;
-    min-width: 400px;
+    position: absolute;
+    width: 80%;
+    max-height: 600px;
     top: 10%;
     left: 50%;
     transform: translateX(-50%);
@@ -196,14 +243,21 @@ header {
 }
 
 .game-list {
-    max-height: 250px;
+    max-height: 400px;
     overflow: scroll;
     border-top: 2px solid gray;
     border-bottom: 2px solid gray;
     box-shadow: 0 0 5px 2px inset darkslategrey;
 }
 
+.game-list-header-row {
+    display: grid;
+    grid-template-columns: 2fr 1fr 1fr 1fr;
+    padding: 0 20px;
+}
+
 .game-list-header {
+    place-self: center;
     font-weight: 600;
     padding-bottom: 10px;
 }
@@ -212,9 +266,10 @@ header {
     text-align: left;
 }
 
-.game-list-row, .game-list-header-row, .game-list-row-nogames {
+.game-list-row, .game-list-row-nogames {
     display: grid;
-    grid-template-columns: 2fr 2fr 1fr;
+    grid-template-columns: 2fr 1fr 1fr 1fr;
+    line-height: 1.2em;
     padding: 0 20px;
 }
 
@@ -232,6 +287,7 @@ header {
 }
 
 .game-list-text {
+    place-self: center;
     pointer-events: none;
 }
