+++ client/gamefiles/html/create-game-modal.html
@@ -1,5 +1,8 @@
-<div class="modal-title">What do you want to call the game?</div>
+<div class="modal-title">New Game</div>
+<div class="subheader">Select Region</div>
+<label for="create-modal-enter-game-name">Enter game name:</label>
 <input id="create-modal-enter-game-name" class="text-line-entry" name="game-name" required minlength="3" maxlength="30" size="30">
+<div id="create-game-regions" class="game-list"></div>
 <div class="error-text">Invalid name</div>
 <div class="wait-text">Please wait...</div>
 <div class="modal-buttons">
