+++ client/gamefiles/html/region-list-row.html
@@ -1,0 +1,7 @@
+<div class="game-list-row">
+    <div class="text-button game-list-text-name" data-set-name=""></div>
+    <div class="set-info">
+        <div class="set-info-planets"></div>
+        <div class="set-info-description"></div>
+    </div>
+</div>
