+++ client/gamefiles/html/user-info-modal.html
@@ -4,6 +4,7 @@
     <div class="game-list-header-row">
         <span id="list-header-gamename" class="game-list-header">Game Name</span>
         <span id="list-header-ownername" class="game-list-header">Game Creator</span>
+        <span id="list-header-region" class="game-list-header">Region</span>
         <span id="list-header-numplayers" class="game-list-header">Number of Players</span>
     </div>
     <div class="game-list"></div>
