+++ client/gamefiles/js/data.js
@@ -1,0 +1,5 @@
+const worlds = {
+    baseGame: ["Earth", "Luna", "Mars", "Venus", "Mercury"]
+};
+
+export default worlds;
