+++ client/gamefiles/js/helpers.js
