+++ client/gamefiles/js/player.js
