+++ client/tools/css/tools.css
@@ -15,6 +15,10 @@ header {
     display: none;
 }
 
+#sign-out {
+    cursor: pointer;
+}
+
 /* navbar */
 
 nav {
