+++ client/tools/html/tools.html
@@ -6,7 +6,7 @@
         <title>Org Tools</title>
         <link rel="stylesheet" type="text/css" href="../css/main.css">
         <link rel="stylesheet" type="text/css" href="../css/tools.css">
-        <link rel="stylesheet" type="text/css" href="../css/firebaseui.css">
+        <link type="text/css" rel="stylesheet" href="https://www.gstatic.com/firebasejs/ui/4.6.0/firebase-ui-auth.css" />
     </head>
     <body>
         <header>Org Tools</header>
@@ -19,6 +19,7 @@
         <nav id="navbar">
             <span id="navlink-card" class="active nav-link">Card Entry</span>
             <span id="navlink-board" class="nav-link">Board Entry</span>
+            <span id="navlink-set" class="nav-link">Set Entry</span>
         </nav>
 
         <main class="tool-container">
@@ -29,6 +30,11 @@
 
             <section id="list-board" class="inactive list-container">
                 <div class="title">Board List</div>
+                <div class="list"></div>
+            </section>
+
+            <section id="list-set" class="inactive list-container">
+                <div class="title">Set List</div>
                 <div class="list"></div>
             </section>
 
@@ -1213,6 +1219,79 @@
                 </form>
             </section>
 
+            <section id="form-container-set" class="inactive form-container">
+                <div class="title">Set Details</div>
+                <form id="form-set">
+                    <table>
+                        <tbody>
+                        <tr>
+                            <td>
+                                <label for="item-name-set">Set</label>
+                                <select id="item-name-set" name="item-name">
+                                    <option value=""></option>
+                                    <option value="innerworlds">The Inner Worlds</option>
+                                    <option value="belt">The Belt</option>
+                                    <option value="galilean">The Galilean Conflict</option>
+                                    <option value="transjovian">The Transjovian War</option>
+                                    <option value="longwar">The Long War</option>
+                                    <option value="kuiper">The Kuiper Belt</option>
+                                </select>
+                            </td>
+                            <td>
+                                <label for="description-set">Description:</label>
+                                <textarea id="description-set" name="description" rows="5" cols="40"></textarea>
+                            </td>
+                            <td>
+                                <label for="region-set">Region</label>
+                                <select id="region-set" name="region" multiple>
+                                    <option value="mercury">Mercury</option>
+                                    <option value="venus">Venus</option>
+                                    <option value="earth">Earth</option>
+                                    <option value="luna">Luna</option>
+                                    <option value="mars">Mars</option>
+                                    <option value="ceres">Ceres</option>
+                                    <option value="vesta">Vesta</option>
+                                    <option value="eunomia">Eunomia</option>
+                                    <option value="exalted">Exalted Temple</option>
+                                    <option value="hildas">Hildas Triangle</option>
+                                    <option value="concordia">Concordia</option>
+                                    <option value="io">Io</option>
+                                    <option value="europa">Europa</option>
+                                    <option value="callisto">Callisto</option>
+                                    <option value="ganymede">Ganymede</option>
+                                    <option value="titan">Titan</option>
+                                    <option value="rhea">Rhea</option>
+                                    <option value="dione">Dione</option>
+                                    <option value="iapetus">Iapetus</option>
+                                    <option value="oberon">Oberon</option>
+                                    <option value="ariel">Ariel</option>
+                                    <option value="miranda">Miranda</option>
+                                    <option value="titania">Titania</option>
+                                    <option value="triton">Triton</option>
+                                    <option value="proteus">Proteus</option>
+                                    <option value="nereid">Nereid</option>
+                                    <option value="neso">Neso</option>
+                                    <option value="pluto">Pluto</option>
+                                    <option value="makemake">Makemake</option>
+                                    <option value="xena">Xena</option>
+                                    <option value="sedna">Sedna</option>
+                                    <option value="haumea">Haumea</option>
+                                    <option value="quorum">The Quorum</option>
+                                    <option value="orcus">Orcus</option>
+                                </select>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td colspan="3" class="button-container">
+                                <input id="submit-board" type="submit" value="Save" class="button-submit">
+                                <input type="reset" value="Clear form" class="button-reset">
+                            </td>
+                        </tr>
+                        </tbody>
+                    </table>
+                </form>
+            </section>
+
             <div id="message" class="shadow">
                 <div class="text"></div>
                 <button class="button-submit">OK</button>
@@ -1220,12 +1299,12 @@
             </div>
         </main>
 
-        <script type="text/javascript" src="../node_modules/firebase/firebase-app.js"></script>
-        <script type="text/javascript" src="../node_modules/firebase/firebase-auth.js"></script>
-        <script type="text/javascript" src="../node_modules/firebase/firebase-database.js"></script>
-        <script type="text/javascript" src="../node_modules/firebaseui/dist/firebaseui.js"></script>
-        <script type="text/javascript" src="../node_modules/jquery/dist/jquery.min.js"></script>
-        <script type="text/javascript" src="../node_modules/jqueryui/jquery-ui.min.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-app.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-auth.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-database.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/ui/4.6.0/firebase-ui-auth.js"></script>
+        <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
+        <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.min.js" integrity="sha256-VazP97ZCwtekAsvgPBSUwPFKdrwD3unUfSGVYrahUqU=" crossorigin="anonymous"></script>
 
         <script type="text/javascript" src="../js/storage.js"></script>
         <script type="text/javascript" src="../js/firebase.js"></script>
