+++ client/tools/index.html
@@ -5,18 +5,19 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <title>Org Tools Login</title>
         <link rel="stylesheet" type="text/css" href="css/main.css">
-        <link rel="stylesheet" type="text/css" href="css/firebaseui.css">
+        <link type="text/css" rel="stylesheet" href="https://www.gstatic.com/firebasejs/ui/4.6.0/firebase-ui-auth.css" />
+
     </head>
     <body>
         <header>Org Tools Login</header>
 
         <div id="firebaseui-auth-container"></div>
 
-        <script type="text/javascript" src="node_modules/firebase/firebase-app.js"></script>
-        <script type="text/javascript" src="node_modules/firebase/firebase-auth.js"></script>
-        <script type="text/javascript" src="node_modules/firebase/firebase-database.js"></script>
-        <script type="text/javascript" src="node_modules/firebaseui/dist/firebaseui.js"></script>
-        <script type="text/javascript" src="node_modules/jquery/dist/jquery.min.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-app.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-auth.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-database.js"></script>
+        <script src="https://www.gstatic.com/firebasejs/ui/4.6.0/firebase-ui-auth.js"></script>
+        <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
 
         <script type="text/javascript" src="js/storage.js"></script>
         <script type="text/javascript" src="js/firebase.js"></script>
