+++ server/modules/data.js
@@ -1,0 +1,5 @@
+const Data = {
+    MAX_PLAYER_COUNT : 5
+};
+
+module.exports = Data;
