+++ server/package.json
@@ -14,11 +14,13 @@
   "main": "public",
   "dependencies": {
     "express": "^4.17.1",
-    "firebase-admin": "^8.3.0",
+    "firebase-admin": "^8.11.0",
     "jqueryui": "^1.11.1",
-    "socket.io": "^2.2.0"
+    "socket.io": "^2.3.0"
   },
   "devDependencies": {
-    "lodash": "^4.17.15"
+    "babel-preset-env": "^1.7.0",
+    "babel-register": "^6.26.0",
+    "lodash": "^4.17.19"
   }
 }
